requirements.txt
python-dotenv
langchain
streamlit
langchain_community